Handle pipeline layout in pipeline state (#70)

+++ include/nox/shader.h
@@ -3,10 +3,12 @@
 #include <nox/export.h>
 #include <nox/shader_types.h>
 
+#include <cstdint>
+
 namespace NOX {
 
 struct ShaderDescriptor {
-    uint8_t stage;
+    uint32_t stage;
 };
 
 class NOX_EXPORT Shader {
